[Fix] #40 - 스캔중 / 스캔완료에 따른 뷰 파일 분리
